Rename and remove tests, consolidate single tests into files with shared concepts
