+++ tests/unit/Image/img_effect.spec.js
@@ -1,5 +1,5 @@
 import { mount } from "@vue/test-utils";
-import CldImage from "../../src/components/CldImage/CldImage.vue";
+import CldImage from "../../../src/components/CldImage/CldImage.vue";
 
 describe("CldImage", () => {
   it("allows transformation as props", () => {
@@ -14,7 +14,7 @@ describe("CldImage", () => {
       components: { CldImage }
     });
     const image = wrapper.find('img');
-    
+
     expect(image.is("img")).toBe(true);
     expect(image.attributes("src")).toEqual(
       `http://res.cloudinary.com/demo/image/upload/e_sepia/face_top`
