+++ tests/unit/Image/img_no_id.spec.js
@@ -1,5 +1,5 @@
 import { mount } from "@vue/test-utils";
-import CldImage from "../../src/components/CldImage/CldImage.vue";
+import CldImage from "../../../src/components/CldImage/CldImage.vue";
 
 describe("CldImage", () => {
   it("should render a src property with an undefined value if a src is not defined", () => {
