+++ tests/unit/Image/img_placeholder_color.spec.js
@@ -1,26 +1,28 @@
 import Vue from "vue";
 import { mount } from "@vue/test-utils";
-import CldImage from "../../src/components/CldImage/CldImage.vue";
+import CldImage from "../../../src/components/CldImage/CldImage.vue";
 
 describe("CldImage", () => {
   describe("handles placeholder attribute", () => {
-    it("pixelate", async () => {
+    it("color", async () => {
       const image = mount({
         template: `
           <cld-image
             cloudName="demo"
             publicId="face_top"
             lazy
-            placeholder="pixelate"
+            placeholder="color"
           />
         `,
         components: { CldImage }
       }).find('img');
       expect(image.is("img")).toBe(true);
       expect(image.attributes("src")).toEqual(
-        `http://res.cloudinary.com/demo/image/upload/e_pixelate:100/face_top`
+        `http://res.cloudinary.com/demo/image/upload/$nh_ih,$nw_iw,c_scale,q_1,w_1/c_scale,h_$nh,w_$nw/face_top`
       );
-      await new Promise(r => Vue.nextTick(r));
+
+      await Vue.nextTick();
+
       expect(image.attributes("src")).toEqual(
         `http://res.cloudinary.com/demo/image/upload/face_top`
       );
