+++ tests/unit/Image/img_placeholder_pixelate.spec.js
@@ -1,24 +1,24 @@
 import Vue from "vue";
 import { mount } from "@vue/test-utils";
-import CldImage from "../../src/components/CldImage/CldImage.vue";
+import CldImage from "../../../src/components/CldImage/CldImage.vue";
 
 describe("CldImage", () => {
   describe("handles placeholder attribute", () => {
-    it("LQ", async () => {
+    it("pixelate", async () => {
       const image = mount({
         template: `
           <cld-image
             cloudName="demo"
             publicId="face_top"
             lazy
-            placeholder="lqip"
+            placeholder="pixelate"
           />
         `,
         components: { CldImage }
       }).find('img');
       expect(image.is("img")).toBe(true);
       expect(image.attributes("src")).toEqual(
-        `http://res.cloudinary.com/demo/image/upload/$nh_ih,$nw_iw,c_scale,q_auto,w_20/c_scale,h_$nh,w_$nw/face_top`
+        `http://res.cloudinary.com/demo/image/upload/e_pixelate:100/face_top`
       );
       await new Promise(r => Vue.nextTick(r));
       expect(image.attributes("src")).toEqual(
