+++ tests/unit/Transformation/img+transform_chained.spec.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import { mount } from "@vue/test-utils";
-import CldTransformation from "../../src/components/CldTransformation/CldTransformation.vue";
-import CldImage from "../../src/components/CldImage/CldImage.vue";
+import CldTransformation from "../../../src/components/CldTransformation/CldTransformation.vue";
+import CldImage from "../../../src/components/CldImage/CldImage.vue";
 
 describe("CldImage with CldTransformation ", () => {
   it("should allow chained transformations", async () => {
